mod/requirements.txt
openai
httpx
requests
python-dotenv
moviepy
pillow
numpy
